refactor: store 패키지 생성 및 Store 파일 이동
