+++ src/main/kotlin/racingcar/store/CarStore.kt
@@ -1,14 +1,16 @@
-package racingcar
+package racingcar.store
+
+import racingcar.Car
 
 object CarStore : Store<Car> {
     private var cars = listOf<Car>()
 
     override fun findAll(): List<Car> {
-        return this.cars
+        return cars
     }
 
     override fun saveAll(list: List<Car>): List<Car> {
-        this.cars = list
-        return this.cars
+        cars = list
+        return cars
     }
 }
